+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added support for tokenizer action when end of input stream is reached
 * Token actions are now always called exactly once for each individual token
+* Tokenizer now supports states
 
 # v0.1.0 (2019-06-27)